Move settings to legacy schema
